Don't iterate forever given an unknown step
This seems to be a problem with #advance, which is used by #every. It
silently returns the same value if options don't contain a known step
key, and that's fine for advance.

But in #every, where we iterate over the range, we end up in an infinite
loop if we don't have a known step key.